refactor: Page 에서 Items component 분리
