+++ src/pages/CartPage.js
@@ -1,21 +1,20 @@
 import React from 'react';
 import { useDispatch, useSelector } from 'react-redux';
 
-import PageTitle from '../../components/PageTitle';
-import FloatingBox from '../../components/FloatingBox';
-import CheckBox from '../../components/utils/CheckBox';
-import Button from '../../components/utils/Button';
-import Flex from '../../components/utils/Flex';
-import PageMessage from '../../components/PageMessage';
+import PageTitle from '../components/PageTitle';
+import CartItem from '../components/CartItem';
+import FloatingBox from '../components/FloatingBox';
+import CheckBox from '../components/utils/CheckBox';
+import Button from '../components/utils/Button';
+import Flex from '../components/utils/Flex';
+import PageMessage from '../components/PageMessage';
 
-import CartItem from './CartItem';
+import { toggleCheckbox, allCheck, allUnCheck } from '../modules/cart';
+import { addPaymentItems } from '../modules/payment';
 
-import { toggleCheckbox, allCheck, allUnCheck } from '../../modules/cart';
-import { addPaymentItems } from '../../modules/payment';
+import { deleteCheckedItems, getTotalPrice } from '../utils';
 
-import { deleteCheckedItems, getTotalPrice } from '../../utils';
-
-import emptyCartImage from '../../asset/empty-cart.png';
+import emptyCartImage from '../asset/empty-cart.png';
 import styled, { css } from 'styled-components';
 
 const CartItemWrapperStyle = css`
